Move helper function to jslib
